Create a Fragment for the MainActivity
Moved everything to a fragment
