+++ app/src/main/java/io/shits/and/gigs/randomcodinglove/RandomLoveDataRepository.kt
@@ -12,5 +12,4 @@ class RandomLoveDataRepository(resources: Resources) : DataAboutAppRepository() 
         addDataToList(resources.getString(R.string.build_type_wrapper, BuildConfig.BUILD_TYPE))
     }
 
-
 }
